src/requirements.txt
Django==4.2
djangorestframework==3.14.0
Markdown==3.5.2
django-filter==24.1
psycopg2==2.9.9
requests==2.31.0
djangorestframework-simplejwt==5.3.1
django-crontab==0.7.1
